archchem

A function to create a data format for chemical analysis datasets in archaeology containing numerical elemental and isotopic data. Loads data from a file (.csv, .xls, .xlsx) or object (dataframe) in R. The data format will contain analytical data as well as corresponding contextual information and metadata (labelled as context).

Usage

as_archchem(
  df,
  id_column = "ID",
  context = c(),
  bdl = c("b.d.", "bd", "b.d.l.", "bdl", "<LOD", "<"),
  bdl_strategy = function() {
     NA_character_
 },
  guess_context_type = TRUE,
  na = c("", "n/a", "NA", "N.A.", "N/A", "na", "-", "n.d.", "n.a.", "#DIV/0!", "#VALUE!",
    "#REF!", "#NAME?", "#NUM!", "#N/A", "#NULL!"),
  drop_columns = FALSE,
  validate = TRUE,
  ...
)

validate(x, quiet = TRUE, ...)

read_archchem(
  path,
  id_column = "ID",
  context = c(),
  delim = ",",
  guess_context_type = TRUE,
  na = c("", "n/a", "NA", "N.A.", "N/A", "na", "-", "n.d.", "n.a.", "#DIV/0!", "#VALUE!",
    "#REF!", "#NAME?", "#NUM!", "#N/A", "#NULL!"),
  bdl = c("b.d.", "bd", "b.d.l.", "bdl", "<LOD", "<"),
  bdl_strategy = function() {
     NA_character_
 },
  drop_columns = FALSE,
  validate = TRUE,
  ...
)

# S3 method for class 'archchem'
format(x, ...)

# S3 method for class 'archchem'
print(x, ...)

remove_units(x, ...)

get_analytical_columns(x)

get_analytical_columns.default(x, ...)

get_contextual_columns(x)

get_contextual_columns.default(x, ...)

Arguments

df

a data.frame containing the input table

id_column

name of the ID column present in df (or in the file at path). Defaults to "ID"

context

Columns that provide contextual (non-measurement) information; may be column names, integer positions, or a logical inclusion vector.

bdl

Strings representing “below detection limit” values. By default, the following are recognized: "b.d.", "bd", "b.d.l.", "bdl", "<LOD", "<".

bdl_strategy

Function used to replace BDL strings. Defaults to a function returning NA.

guess_context_type

If TRUE, attempt to infer appropriate classes for context columns.

na

Character vector of strings to be interpret as missing values.

drop_columns

...

validate

...

...

Additional arguments passed to the respective import functions. See their documentation for details:

readxl::read_excel() for file formats .xlsx or .xls

readr::read_delim() for all other file formats.

x

an object of class archchem

quiet

...

path

path to the file that should be read

delim

A character string with the separator for tabular data. Must be provided for all file types except .xlsx or .xls. Default to ,. Use \t for tab-separated data.

Value

Returns a data structure archchem which is a tibble derived-object

The function prints a short summary about the dataset, including a list of all context columns and analytical data columns.

Details

The data files can be fairly freeform, i.e. no specified elements, oxides, or isotopic ratios are required and no exact order of these needs to be adhered to. Analyses can contain as many analytical columns as necessary.

The column that contains the unique samples identification is specified using the ID argument. If the dataset contains duplicate ids, the following warning will return: Detected multiple data rows with the same ‘ID’. They will be renamed consecutively using the following convention: _1,_2, ... _n

Metadata contained within the dataset must be specified using the context argument. If any column in the dataframe is not specified as context and not recognised as an analytical column, this will result in the error: Column name could not be parsed. Either analytical columns do not conform to ASTR conventions or contextual columns are not specified as such.

Below detection limit notation (i.e. ‘b.d.’, ‘bd’, ‘b.d.l.’, ‘bdl’, ‘<LOD’, or ‘<..’) for element and oxide concentrations is specified using the bdl argument. One or more notations can be used as is appropriate for the dataset, and can be notations not included in the list above. The argument bdl_strategy is used to specify the value for handling detection limits. This is to facilitate the different handling needs of the detection limit for future statistical applications, as opposed to automatically assigning such values as ‘NA’.

Missing values (NA) are allowed anywhere in the data file body, and those found in an analytical data column will be replaced by NA automatically.