Decompose an R function

Decompose an R function

Usage

ast_decompose(x)

Arguments

x

a function

Value

result of utils::getParseData(), a data.frame; but we wrap it in a thin S3 wrapper to make it easy to see how the function changes as you modify it with ast_modify()

Examples

foo <- function(x) {
  x + 1
}
foo(5)
#> [1] 6
df <- ast_decompose(foo)
df
#> { fxn: foo }
#> function (x)
#> {
#>     x + 1
#> }
data.frame(df)
#>    line1 col1 line2 col2 id parent          token terminal     text
#> 1      1    1     1    8  1     23       FUNCTION     TRUE function
#> 2      1   10     1   10  2     23            '('     TRUE        (
#> 3      1   11     1   11  3     23 SYMBOL_FORMALS     TRUE        x
#> 4      1   12     1   12  4     23            ')'     TRUE        )
#> 7      2    1     2    1  7     20            '{'     TRUE        {
#> 9      3    5     3    5  9     11         SYMBOL     TRUE        x
#> 10     3    7     3    7 10     15            '+'     TRUE        +
#> 12     3    9     3    9 12     13      NUM_CONST     TRUE        1
#> 18     4    1     4    1 18     20            '}'     TRUE        }
attributes(df)
#> $names
#> [1] "line1"    "col1"     "line2"    "col2"     "id"       "parent"   "token"   
#> [8] "terminal" "text"    
#> 
#> $srcfile
#> <text> 
#> 
#> $row.names
#> [1]  1  2  3  4  7  9 10 12 18
#> 
#> $class
#> [1] "ast"        "data.frame"
#> 
#> $name
#> [1] "foo"
#> 
#> $expr
#>    line1 col1 line2 col2 id parent token terminal
#> 23     1    1     4    1 23      0  expr    FALSE
#> 20     2    1     4    1 20     23  expr    FALSE
#> 15     3    5     3    9 15     20  expr    FALSE
#> 11     3    5     3    5 11     15  expr    FALSE
#> 13     3    9     3    9 13     15  expr    FALSE
#>                              text
#> 23 function (x) \n{\n    x + 1\n}
#> 20                {\n    x + 1\n}
#> 15                          x + 1
#> 11                              x
#> 13                              1
#> 
attr(df, "expr")
#>    line1 col1 line2 col2 id parent token terminal
#> 23     1    1     4    1 23      0  expr    FALSE
#> 20     2    1     4    1 20     23  expr    FALSE
#> 15     3    5     3    9 15     20  expr    FALSE
#> 11     3    5     3    5 11     15  expr    FALSE
#> 13     3    9     3    9 13     15  expr    FALSE
#>                              text
#> 23 function (x) \n{\n    x + 1\n}
#> 20                {\n    x + 1\n}
#> 15                          x + 1
#> 11                              x
#> 13                              1
attr(df, "expr")[1,"text"] # the original fxn
#> [1] "function (x) \n{\n    x + 1\n}"