Modify an R function

Modify an R function

ast_modify(x, from, to, if_many = "random", no_match = stop)

Arguments

x
an object of class ast

from
(character) character string to replace. note that we look for an exact match

to
(character) character string to put in place of from

if_many
(character) if multiple matches to the from parameter input, should we randomly select one to replace, replace the first instance, or replace all? one of: first, all, random

no_match
(function) how to deal with no matches. by default we stop(), but you can set to warning() or message()

Value

same as the input, an object of class ast, but modified

Details

we check that the from input has a match in the function data, if not, we fail out

Examples

foo <- function(x) {
  x + 1
}
foo(5)

#> [1] 6

# decompose the function
df <- ast_decompose(foo)
df

#> { fxn: foo }
#> function (x)
#> {
#>     x + 1
#> }

data.frame(df)

#>    line1 col1 line2 col2 id parent          token terminal     text
#> 1      1    1     1    8  1     23       FUNCTION     TRUE function
#> 2      1   10     1   10  2     23            '('     TRUE        (
#> 3      1   11     1   11  3     23 SYMBOL_FORMALS     TRUE        x
#> 4      1   12     1   12  4     23            ')'     TRUE        )
#> 7      2    1     2    1  7     20            '{'     TRUE        {
#> 9      3    5     3    5  9     11         SYMBOL     TRUE        x
#> 10     3    7     3    7 10     15            '+'     TRUE        +
#> 12     3    9     3    9 12     13      NUM_CONST     TRUE        1
#> 18     4    1     4    1 18     20            '}'     TRUE        }

attr(df, "expr")

#>    line1 col1 line2 col2 id parent token terminal
#> 23     1    1     4    1 23      0  expr    FALSE
#> 20     2    1     4    1 20     23  expr    FALSE
#> 15     3    5     3    9 15     20  expr    FALSE
#> 11     3    5     3    5 11     15  expr    FALSE
#> 13     3    9     3    9 13     15  expr    FALSE
#>                              text
#> 23 function (x) \n{\n    x + 1\n}
#> 20                {\n    x + 1\n}
#> 15                          x + 1
#> 11                              x
#> 13                              1

# modify an aspect of the function
out <- ast_modify(x = df, from = "+", to = "-")
out

#> { fxn: foo }
#> function (x)
#> {
#>     x - 1
#> }

class(out)

#> [1] "ast"        "data.frame"

attributes(out)

#> $names
#>  [1] "line1"           "col1"            "line2"           "col2"           
#>  [5] "id"              "parent"          "token"           "terminal"       
#>  [9] "text"            "mutated"         "mutated_from_to"
#> 
#> $srcfile
#> <text> 
#> 
#> $row.names
#> [1]  1  2  3  4  7  9 10 12 18
#> 
#> $name
#> [1] "foo"
#> 
#> $expr
#>    line1 col1 line2 col2 id parent token terminal
#> 23     1    1     4    1 23      0  expr    FALSE
#> 20     2    1     4    1 20     23  expr    FALSE
#> 15     3    5     3    9 15     20  expr    FALSE
#> 11     3    5     3    5 11     15  expr    FALSE
#> 13     3    9     3    9 13     15  expr    FALSE
#>                              text
#> 23 function (x) \n{\n    x + 1\n}
#> 20                {\n    x + 1\n}
#> 15                          x + 1
#> 11                              x
#> 13                              1
#> 
#> $class
#> [1] "ast"        "data.frame"
#> 

data.frame(out)

#>    line1 col1 line2 col2 id parent          token terminal     text mutated
#> 1      1    1     1    8  1     23       FUNCTION     TRUE function   FALSE
#> 2      1   10     1   10  2     23            '('     TRUE        (   FALSE
#> 3      1   11     1   11  3     23 SYMBOL_FORMALS     TRUE        x   FALSE
#> 4      1   12     1   12  4     23            ')'     TRUE        )   FALSE
#> 7      2    1     2    1  7     20            '{'     TRUE        {   FALSE
#> 9      3    5     3    5  9     11         SYMBOL     TRUE        x   FALSE
#> 10     3    7     3    7 10     15            '+'     TRUE        -    TRUE
#> 12     3    9     3    9 12     13      NUM_CONST     TRUE        1   FALSE
#> 18     4    1     4    1 18     20            '}'     TRUE        }   FALSE
#>    mutated_from_to
#> 1             <NA>
#> 2             <NA>
#> 3             <NA>
#> 4             <NA>
#> 7             <NA>
#> 9             <NA>
#> 10             +,-
#> 12            <NA>
#> 18            <NA>

attr(out, "expr")

#>    line1 col1 line2 col2 id parent token terminal
#> 23     1    1     4    1 23      0  expr    FALSE
#> 20     2    1     4    1 20     23  expr    FALSE
#> 15     3    5     3    9 15     20  expr    FALSE
#> 11     3    5     3    5 11     15  expr    FALSE
#> 13     3    9     3    9 13     15  expr    FALSE
#>                              text
#> 23 function (x) \n{\n    x + 1\n}
#> 20                {\n    x + 1\n}
#> 15                          x + 1
#> 11                              x
#> 13                              1

# more examples
bar <- function(x) x / 6
(z <- ast_decompose(bar))

#> { fxn: bar }
#> function (x)
#> x/6

ast_modify(z, from = "/", to = "*")

#> { fxn: bar }
#> function (x)
#> x*6

# to get the new function, pass through ast_recompose
b <- ast_modify(z, from = "/", to = "*")
newbar <- ast_recompose(b, TRUE)
bar(7)

#> [1] 1.166667

eval(newbar)(7)

#> [1] 42

# multiple from matches
foo <- function(x) {
  w <- x + 1
  w + 5
}
foo(1)

#> [1] 7

x <- ast_decompose(foo)
(w <- ast_modify(x, "+", "-"))

#> { fxn: foo }
#> function (x)
#> {
#>     w <- x - 1
#>     w + 5
#> }

eval(ast_recompose(w, TRUE))(1)

#> [1] 5

# if_many options
ast_modify(x, "+", "-", if_many = "random")

#> { fxn: foo }
#> function (x)
#> {
#>     w <- x - 1
#>     w + 5
#> }

ast_modify(x, "+", "-", if_many = "random")

#> { fxn: foo }
#> function (x)
#> {
#>     w <- x + 1
#>     w - 5
#> }

ast_modify(x, "+", "-", if_many = "random")

#> { fxn: foo }
#> function (x)
#> {
#>     w <- x - 1
#>     w + 5
#> }

ast_modify(x, "+", "-", if_many = "first")

#> { fxn: foo }
#> function (x)
#> {
#>     w <- x - 1
#>     w + 5
#> }

ast_modify(x, "+", "-", if_many = "all")

#> { fxn: foo }
#> function (x)
#> {
#>     w <- x - 1
#>     w - 5
#> }